+++ CHANGELOG.md
@@ -7,6 +7,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.155](https://github.com/release-plz/release-plz/compare/release-plz-v0.3.154...release-plz-v0.3.155) - 2026-02-05
+
+### Fixed
+
+- correctly recognize when crate already published ([#2649](https://github.com/release-plz/release-plz/pull/2649))
+- read cargo registry token from env var correctly ([#2647](https://github.com/release-plz/release-plz/pull/2647))
+
+### Other
+
+- use cargo env variable for registry token ([#2637](https://github.com/release-plz/release-plz/pull/2637))
+- update dependencies ([#2646](https://github.com/release-plz/release-plz/pull/2646))
+
+## [0.3.154](https://github.com/release-plz/release-plz/compare/release-plz-v0.3.153...release-plz-v0.3.154) - 2026-02-03
+
+### Fixed
+
+- git-only works with local dependencies ([#2629](https://github.com/release-plz/release-plz/pull/2629))
+
+### Other
+
+- improve git-only workspace test ([#2635](https://github.com/release-plz/release-plz/pull/2635))
+- test cargo lock update ([#2632](https://github.com/release-plz/release-plz/pull/2632))
+- *(tests)* show release-plz stdout better ([#2630](https://github.com/release-plz/release-plz/pull/2630))
+- remove release_plz_core tests ([#2633](https://github.com/release-plz/release-plz/pull/2633))
+
 ## [0.3.153](https://github.com/release-plz/release-plz/compare/release-plz-v0.3.152...release-plz-v0.3.153) - 2026-01-24
 
 ### Other